// frontend/packages/common/chat-area/chat-area/src/store/types.ts
interface ExtraMessageFields {
  // Internal properties, only used to send failed scenarios
  _sendFailed?: boolean;
  // Internal properties for marking history
  _fromHistory?: boolean;
  // Internal property to flag the message as coming from onboarding
  _fromOnboarding?: boolean;
}
/* eslint-enable @typescript-eslint/naming-convention -- explain what... */

// eslint-disable-next-line @typescript-eslint/naming-convention -- non-compliance
type _Message<T extends ContentType, V = unknown> = OriginMessage<T, V> &
  ExtraMessageFields;

export type Message<
  T extends ContentType = ContentType,
  V = unknown,
> = _Message<T, V> & ExtraMessageFields;

export interface MessageIdStruct {
  message_id: string;
  extra_info: {
    local_message_id: string;
  };
}
// ...
export interface MessageExtraInfoBotState {
  bot_id?: string;
  agent_id?: string;
  agent_name?: string;
  agent_type?: AgentType;
  awaiting?: string;
}
// ...
export interface MessageMeta
  extends MessageIdStruct,
    Pick<ExtraMessageFields, '_fromHistory'> {
  /** Message display Copy, regenerate and other buttons */
  showActions: boolean;
  /** The agent split line is displayed after the message. */
  showMultiAgentDivider: boolean;
  /** User question is being sent. */
  isSending: boolean;
  /** Message sending or receiving failed */
  isFail: boolean;
  /** Message receiving */
  isReceiving: boolean;
  /** Exhibits function calls only when true; terminated by the corresponding tool_response;
   * Note that the non-debug area (home, store, web sdk) uses this value to determine the display of the function call; the debug area has other judgment logic */
  isFunctionCalling: boolean;
  /** Whether to show suggestion: Only show the last reply when there is no dividing line */
  // TODO: lsy confirm whether to delete
  // showSuggestions: boolean;
  /** Is it a message from the last message group? */
  isFromLatestGroup: boolean;
  /** Contains only message.type === 'answer'  */
  isGroupFirstAnswer: boolean;
  /** Contains answer, [function_call, verbose], query; in decreasing order, function_call and verbose in variable order */
  isGroupLastMessage: boolean;
  /** The last message of the current group. Type === 'answer' */
  isGroupLastAnswerMessage: boolean;
  /**
   * Whether to hide the avatar
   */
  hideAvatar: boolean;
  role: Message['role'];
  type: Message['type'];
  sectionId: string;
  replyId: string;
  botState: MessageExtraInfoBotState;
  /**
   * Whether there is an agent jump verbose message in the current group, which is used to determine whether to display the agent split line
   */
  beforeHasJumpVerbose: boolean;
  verboseMsgType: VerboseMsgType | '';
  source: MessageSource | undefined;
  // Is the card disabled?
  cardDisabled: boolean;
}